Refactor login and user check functions
This commit refactors the login page component and adds a new function checkUser() in the common folder. The checkUser() function is responsible for checking if the user is logged in and retrieving user data from the server. It throws an error if the user is not found or if the user data is not available. The login.ts file is also added, which contains the signIn() function responsible for authenticating the user with email and password.

Refactor login.ts and checkUser.ts files to improve code organization and reusability.